Move `ImageAnalysis.Analyzer` to `scanid.qrcodeanalyzer` package
