+++ app/src/main/java/org/simple/clinic/scanid/ScanSimpleIdScreen.kt
@@ -44,8 +44,8 @@ import org.simple.clinic.util.unsafeLazy
 import org.simple.clinic.widgets.UiEvent
 import org.simple.clinic.widgets.hideKeyboard
 import org.simple.clinic.widgets.qrcodescanner.BitmapUtils
-import org.simple.clinic.widgets.qrcodescanner.MLKitQrCodeAnalyzer
-import org.simple.clinic.widgets.qrcodescanner.ZxingQrCodeAnalyzer
+import org.simple.clinic.scanid.qrcodeanalyzer.MLKitQrCodeAnalyzer
+import org.simple.clinic.scanid.qrcodeanalyzer.ZxingQrCodeAnalyzer
 import java.util.concurrent.Executors
 import javax.inject.Inject
 import kotlin.math.abs
