+++ app/src/main/java/org/simple/clinic/scanid/qrcodeanalyzer/MLKitQrCodeAnalyzer.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.widgets.qrcodescanner
+package org.simple.clinic.scanid.qrcodeanalyzer
 
 import android.annotation.SuppressLint
 import androidx.camera.core.ImageAnalysis
@@ -9,6 +9,7 @@ import com.google.mlkit.vision.barcode.Barcode
 import com.google.mlkit.vision.barcode.BarcodeScannerOptions
 import com.google.mlkit.vision.barcode.BarcodeScanning
 import com.google.mlkit.vision.common.InputImage
+import org.simple.clinic.widgets.qrcodescanner.BitmapUtils
 
 class MLKitQrCodeAnalyzer(
     private val bitmapUtils: BitmapUtils,
