+++ app/src/main/java/org/simple/clinic/scanid/qrcodeanalyzer/OnQrCodeDetected.kt
@@ -1,0 +1,3 @@
+package org.simple.clinic.scanid.qrcodeanalyzer
+
+typealias OnQrCodeDetected = (String) -> Unit
