--- app/src/main/java/org/simple/clinic/widgets/qrcodescanner/OnQrCodeDetected.kt
@@ -1,3 +1,0 @@
-package org.simple.clinic.widgets.qrcodescanner
-
-typealias OnQrCodeDetected = (String) -> Unit
